requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-login==0.6.2
gunicorn==21.2.0
python-dotenv==1.0.0
pandas==2.1.1
matplotlib==3.8.0
plotly==5.17.0
yfinance==0.2.31
requests==2.31.0
Werkzeug==2.3.7
openai
Flask-Migrate
python-dotenv
fpdf
newsapi-python
